+++ scripts/format/format.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const pfs = require('fs').promises;
+const fsp = require('fs').promises;
 const getOutputJobsFromIndex = require('./get-output-jobs-from-index');
 const runOutputJob = require('./run-output-job');
 const getIndexFromOutputSamples = require('./get-index-from-output-samples');
@@ -33,7 +33,7 @@ const prerenderedOutputJobs = getOutputJobsFromIndex(
 
 const outputJobs = baseOutputJobs.concat(prerenderedOutputJobs);
 
-pfs
+fsp
   .readFile(outputIndexPath)
   .then(
     data => JSON.parse(data),
@@ -52,22 +52,57 @@ pfs
     sequentialPromises(outputFns, 'Creating Samples', 'samples created').then(
       outputSamples => {
         const appendIndex = getIndexFromOutputSamples(outputSamples);
-        const outputIndex = {
-          wav: {},
-          mp3: {},
-          ogg: {},
-        };
-        Object.keys(outputIndex).forEach(format => {
-          outputIndex[format] = Object.assign(
-            {},
-            existingOutputIndex[format],
-            appendIndex[format]
-          );
-        });
-        return pfs.writeFile(
-          './dist/index.json',
-          JSON.stringify(outputIndex, null),
-          'utf8'
+        const outputIndex = Object.keys(appendIndex).reduce(
+          (byInstrumentName, instrumentName) => {
+            const existingInstrumentIndex = existingOutputIndex[instrumentName]
+              ? existingOutputIndex[instrumentName]
+              : {};
+            const formats = Array.from(
+              new Set(
+                Object.keys(existingInstrumentIndex).concat(
+                  Object.keys(appendIndex[instrumentName])
+                )
+              )
+            );
+            byInstrumentName[instrumentName] = formats.reduce(
+              (byFormat, format) => {
+                byFormat[format] = Object.assign(
+                  {},
+                  existingInstrumentIndex[format],
+                  appendIndex[format]
+                );
+                return byFormat;
+              },
+              {}
+            );
+            return byInstrumentName;
+          },
+          {}
+        );
+        const singleFormatIndicies = Object.keys(outputIndex).reduce(
+          (byFormat, instrumentName) => {
+            const instrumentFilesByFormat = outputIndex[instrumentName];
+            Object.keys(instrumentFilesByFormat).forEach(format => {
+              if (!byFormat[format]) {
+                byFormat[format] = {};
+              }
+              byFormat[format][instrumentName] =
+                instrumentFilesByFormat[format];
+            });
+            return byFormat;
+          },
+          {}
+        );
+        const indexFiles = Object.keys(singleFormatIndicies)
+          .map(format => [
+            `./dist/${format}.json`,
+            singleFormatIndicies[format],
+          ])
+          .concat([['./dist/index.json', outputIndex]]);
+        return Promise.all(
+          indexFiles.map((filename, data) =>
+            fsp.writeFile(filename, JSON.stringify(data), 'utf8')
+          )
         );
       }
     );