+++ src/get-full-path.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const getFullPath = ({ host, instrumentName, format, basename }) =>
+  `${[host, instrumentName, format, basename].join('/')}.${format}`;
+
+module.exports = getFullPath;